backend/requirements.txt
Flask==2.2.5
Flask-Cors==3.0.10
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.2
joblib==1.4.2
opencv-python==4.8.1.78
Pillow==10.0.1
